@@ -14,7 +14,7 @@
  * ---  END LICENSE_HEADER BLOCK  ---
 */
 
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
 import Axios from 'axios';
 import './collections/Collection.scss';
 import CollectionListStickyUtils from './collections/list/CollectionListStickyUtils';
@@ -24,52 +24,33 @@ import CollectionsFilterNoResults from './collections/CollectionsFilterNoResults
 import PropTypes from 'prop-types';
 import LoadingSpinner from './ui/LoadingSpinner';
 
-class CollectionList extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      filter: props.filter ? props.filter : '',
-      searchResult: [],
-      filteredResult: [],
-      maxItems: 4,
-      sort: 'unit',
-      isLoading: false
+const CollectionList = ({ baseUrl, filter: initialFilter = '' }) => {
+  const [filter, setFilter] = useState(initialFilter);
+  const [searchResult, setSearchResult] = useState([]);
+  const [filteredResult, setFilteredResult] = useState([]);
+  const [maxItems] = useState(4);
+  const [sort, setSort] = useState('unit');
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(() => {
+    const retrieveResults = async () => {
+      setIsLoading(true);
+      try {
+        const response = await Axios({ url: baseUrl });
+        const results = response.data;
+        setSearchResult(results);
+        setFilteredResult(filterCollections(filter, results));
+        setIsLoading(false);
+      } catch (error) {
+        console.log('Error in retrieveResults(): ', error);
+        setIsLoading(false);
+      }
     };
-  }
 
-  componentDidMount() {
-    this.retrieveResults();
-  }
+    retrieveResults();
+  }, [baseUrl]);
 
-  componentDidUpdate(prevProps, prevState) {
-    if (prevState.filter != this.state.filter) {
-      this.setState({
-        filteredResult: this.filterCollections(
-          this.state.filter,
-          this.state.searchResult
-        )
-      });
-    }
-  }
-
-  async retrieveResults() {
-    this.setState({ isLoading: true });
-    let url = this.props.baseUrl;
-
-    try {
-      const response = await Axios({ url });
-      this.setState({
-        searchResult: response.data,
-        filteredResult: this.filterCollections(this.state.filter, response.data),
-        isLoading: false
-      });
-    } catch (error) {
-      console.log('Error in retrieveResults(): ', error);
-      Promise.resolve([]);
-    }
-  }
-
-  sortByAZ(list) {
+  const sortByAZ = (list) => {
     let sortedArray = list.slice();
     sortedArray.sort((col1, col2) => {
       if (col1.name < col2.name) {
@@ -81,9 +62,9 @@ class CollectionList extends Component {
       return 0;
     });
     return sortedArray;
-  }
+  };
 
-  filterCollections(filter, collections) {
+  const filterCollections = (filter, collections) => {
     let filteredArray = [];
     let downcaseFilter = filter.toLowerCase();
     collections.forEach(col => {
@@ -97,53 +78,169 @@ class CollectionList extends Component {
       }
     });
     return filteredArray;
-  }
-
-  handleFilterChange = event => {
-    this.setState({ filter: event.target.value });
   };
 
-  handleSortChange = value => {
-    this.setState({ sort: value });
+  const handleFilterChange = (event) => {
+    setFilter(event.target.value);
   };
 
-  handleSubmit = event => {
+  const handleSortChange = (value) => {
+    setSort(value);
+  };
+
+  const handleSubmit = (event) => {
     event.preventDefault();
-  }
+  };
 
-  render() {
-    const { filter, sort, filteredResult = [], maxItems, isLoading } = this.state;
+  return (
+    <>
+      <CollectionListStickyUtils
+        filter={filter}
+        handleFilterChange={handleFilterChange}
+        sort={sort}
+        handleSortChange={handleSortChange}
+        handleSubmit={handleSubmit}
+      />
+      {isLoading && <LoadingSpinner isLoading={isLoading} />}
+      {(filteredResult.length === 0 && !isLoading) && <CollectionsFilterNoResults />}
 
-    return (
-      <div>
-        <CollectionListStickyUtils
-          filter={filter}
-          handleFilterChange={this.handleFilterChange}
-          sort={sort}
-          handleSortChange={this.handleSortChange}
-	  handleSubmit={this.handleSubmit}
-        />
-        {isLoading && <LoadingSpinner isLoading={isLoading} />}
-        {(filteredResult.length === 0 && !isLoading) && <CollectionsFilterNoResults />}
-
-        <div className="collection-list">
-          {sort === 'az' ? (
-            <CollectionsSortedByAZ
-              filteredResult={filteredResult}
-              sortByAZ={this.sortByAZ}
-            />
-          ) : (
-            <CollectionsSortedByUnit
-              filteredResult={filteredResult}
-              sortByAZ={this.sortByAZ}
-              maxItems={maxItems}
-            />
-          )}
-        </div>
+      <div className="collection-list">
+        {sort === 'az' ? (
+          <CollectionsSortedByAZ
+            filteredResult={filteredResult}
+            sortByAZ={sortByAZ}
+          />
+        ) : (
+          <CollectionsSortedByUnit
+            filteredResult={filteredResult}
+            sortByAZ={sortByAZ}
+            maxItems={maxItems}
+          />
+        )}
       </div>
-    );
-  }
-}
+    </>
+  );
+};
+// class CollectionList extends Component {
+//   constructor(props) {
+//     super(props);
+//     this.state = {
+//       filter: props.filter ? props.filter : '',
+//       searchResult: [],
+//       filteredResult: [],
+//       maxItems: 4,
+//       sort: 'unit',
+//       isLoading: false
+//     };
+//   }
+
+//   componentDidMount() {
+//     this.retrieveResults();
+//   }
+
+//   componentDidUpdate(prevProps, prevState) {
+//     if (prevState.filter != this.state.filter) {
+//       this.setState({
+//         filteredResult: this.filterCollections(
+//           this.state.filter,
+//           this.state.searchResult
+//         )
+//       });
+//     }
+//   }
+
+//   async retrieveResults() {
+//     this.setState({ isLoading: true });
+//     let url = this.props.baseUrl;
+
+//     try {
+//       const response = await Axios({ url });
+//       this.setState({
+//         searchResult: response.data,
+//         filteredResult: this.filterCollections(this.state.filter, response.data),
+//         isLoading: false
+//       });
+//     } catch (error) {
+//       console.log('Error in retrieveResults(): ', error);
+//       Promise.resolve([]);
+//     }
+//   }
+
+//   sortByAZ(list) {
+//     let sortedArray = list.slice();
+//     sortedArray.sort((col1, col2) => {
+//       if (col1.name < col2.name) {
+//         return -1;
+//       }
+//       if (col1.name > col2.name) {
+//         return 1;
+//       }
+//       return 0;
+//     });
+//     return sortedArray;
+//   }
+
+//   filterCollections(filter, collections) {
+//     let filteredArray = [];
+//     let downcaseFilter = filter.toLowerCase();
+//     collections.forEach(col => {
+//       if (
+//         col.name.toLowerCase().includes(downcaseFilter) ||
+//         col.unit.toLowerCase().includes(downcaseFilter) ||
+//         (col.description &&
+//           col.description.toLowerCase().includes(downcaseFilter))
+//       ) {
+//         filteredArray.push(col);
+//       }
+//     });
+//     return filteredArray;
+//   }
+
+//   handleFilterChange = event => {
+//     this.setState({ filter: event.target.value });
+//   };
+
+//   handleSortChange = value => {
+//     this.setState({ sort: value });
+//   };
+
+//   handleSubmit = event => {
+//     event.preventDefault();
+//   };
+
+//   render() {
+//     const { filter, sort, filteredResult = [], maxItems, isLoading } = this.state;
+
+//     return (
+//       <div>
+//         <CollectionListStickyUtils
+//           filter={filter}
+//           handleFilterChange={this.handleFilterChange}
+//           sort={sort}
+//           handleSortChange={this.handleSortChange}
+//           handleSubmit={this.handleSubmit}
+//         />
+//         {isLoading && <LoadingSpinner isLoading={isLoading} />}
+//         {(filteredResult.length === 0 && !isLoading) && <CollectionsFilterNoResults />}
+
+//         <div className="collection-list">
+//           {sort === 'az' ? (
+//             <CollectionsSortedByAZ
+//               filteredResult={filteredResult}
+//               sortByAZ={this.sortByAZ}
+//             />
+//           ) : (
+//             <CollectionsSortedByUnit
+//               filteredResult={filteredResult}
+//               sortByAZ={this.sortByAZ}
+//               maxItems={maxItems}
+//             />
+//           )}
+//         </div>
+//       </div>
+//     );
+//   }
+// }
 
 CollectionList.propTypes = {
   baseUrl: PropTypes.string,
