@@ -15,6 +15,7 @@
 */
 
 import React from 'react';
+import { createRoot } from "react-dom/client";
 import { render } from 'react-dom';
 import Search from './Search';
 
@@ -23,4 +24,6 @@ const props = {
   collection: 'Chris Test'
 };
 
-render(<Search {...props} />, document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
+root.render(<Search {...props} />);
+// render(<Search {...props} />, document.getElementById('root'));
